refactor: move auth folder into cloud folder
