+++ apps/cli/src/cloud/clients.tsx
@@ -1,4 +1,4 @@
-import { readCredentialsFile } from '@/auth/storage'
+import { readCredentialsFile } from '@/cloud/auth/storage'
 import { Failure } from '@/components/failure'
 import type { LocalContext } from '@/context'
 import { env } from '@/env'
